refactor(server): extract memory preflight policy
